# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's BASE",
        "description": "API's for transportation management",
        "version": "1.0.0"
    },
    "paths": {
...
        "/base-backend/public/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Obtener información con filtros y ordenamiento",
                "operationId": "18aafefcfc221771d1d291a52641e81b",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Filtrar por username",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre de usuario",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "person.name",
                        "in": "query",
                        "description": "Filtrar por nombre de la persona",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "person.business_name",
                        "in": "query",
                        "description": "Filtrar por nombre de la empresa",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Fecha de inicio",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Fecha de fin",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de Empresas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Crear usuario",
                "operationId": "3ba0c1c787e25768d2cb3d49af1c0ca3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password",
                                    "number_document"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "juan.perez@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "type_document": {
                                        "type": "string",
                                        "example": "DNI"
                                    },
                                    "type_person": {
                                        "type": "string",
                                        "example": "natural"
                                    },
                                    "names": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "father_surname": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "mother_surname": {
                                        "type": "string",
                                        "example": "Gómez"
                                    },
                                    "business_name": {
                                        "type": "string",
                                        "example": "Pérez Enterprises"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle Ficticia 123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+51987654321"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan.perez@example.com"
                                    },
                                    "number_document": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "ocupation": {
                                        "type": "string",
                                        "example": "Trabajador"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "La persona ya tiene un usuario asociado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/base-backend/public/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Obtener detalles de un usuario por ID",
                "operationId": "b4b22a6e7533fa50451edc5a3de5dca2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del Usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario No Encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario No Encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Actualizar usuario",
                "operationId": "1d9a94396f2a71b3ec82bb44612ce4b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password",
                                    "number_document"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "juan.perez@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "type_document": {
                                        "type": "string",
                                        "example": "DNI"
                                    },
                                    "type_person": {
                                        "type": "string",
                                        "example": "natural"
                                    },
                                    "names": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "father_surname": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "mother_surname": {
                                        "type": "string",
                                        "example": "Gómez"
                                    },
                                    "business_name": {
                                        "type": "string",
                                        "example": "Pérez Enterprises"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle Ficticia 123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+51987654321"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan.perez@example.com"
                                    },
                                    "number_document": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "ocupation": {
                                        "type": "string",
                                        "example": "Trabajador"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Datos inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Eliminar persona por ID",
                "operationId": "6a7e4965d8d999867fa1b0ef3f528bc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de Usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "persona eliminada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario No Encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario No Encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Permission": {
                "title": "Permission",
                "description": "Permission model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "users"
                    },
                    "type": {
                        "type": "string",
                        "example": "Tipo-01"
                    },
                    "status": {
                        "type": "string",
                        "example": "Activo"
                    },
                    "person_id": {
                        "description": "Person Id",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
            "Person": {
                "title": "Person",
                "description": "Person model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type_document": {
                        "type": "string",
                        "example": "DNI"
                    },
                    "type_person": {
                        "type": "string",
                        "example": "Individual"
                    },
                    "number_document": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "names": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "father_surname": {
                        "type": "string",
                        "example": "Perez"
                    },
                    "mother_surname": {
                        "type": "string",
                        "example": "Lopez"
                    },
                    "business_name": {
                        "type": "string",
                        "example": "Perez & Co."
                    },
                    "address": {
                        "type": "string",
                        "example": "Av. Siempre Viva 123"
                    },
                    "phone": {
                        "type": "string",
                        "example": "987654321"
                    },
                    "email": {
                        "type": "string",
                        "example": "juanperez@example.com"
                    },
                    "occupation": {
                        "type": "string",
                        "example": "Engineer"
                    },
                    "status": {
                        "type": "string",
                        "example": "Activo"
                    },
                    "server_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-02T12:00:00Z"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null
                    }
                },
                "type": "object"
            },
            "Rol": {
                "title": "Rol",
                "description": "Rol model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Cajero"
                    },
                    "person_id": {
                        "description": "Person Id",
                        "type": "integer",
                        "example": "1"
                    },
                    "permissions": {
                        "description": "Lista de permisos asociados al rol",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Permission"
                        }
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "email": {
                        "type": "string",
                        "example": "miguel@gmail.com"
                    },
                    "person_id": {
                        "description": "Person Id",
                        "type": "integer",
                        "example": "1"
                    },
                    "person": {
                        "$ref": "#/components/schemas/Person"
                    },
                    "rol_id": {
                        "description": "Rol Id",
                        "type": "integer",
                        "example": "1"
                    },
                    "rol": {
                        "$ref": "#/components/schemas/Rol"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}